refactor: Reorganize Folder Structure; Few Css Tweaks
