+++ app/components/HomePage/MediaRankingSection/index.tsx
@@ -1,8 +1,8 @@
 "use client"
 import React, { useEffect, useState } from 'react'
 import styles from "./component.module.css"
-import MediaListCoverInfo from '../MediaItemCoverInfo2'
-import NavButtons from '../NavButtons'
+import MediaListCoverInfo from '../../MediaItemCoverInfo2'
+import NavButtons from '../../NavButtons'
 import { ApiDefaultResult } from '@/app/ts/interfaces/apiAnilistDataInterface'
 import API from "@/api/anilist"
