+++ app/components/HomePage/NewestMediaSection/index.tsx
@@ -1,9 +1,9 @@
 "use client"
 import React, { useEffect, useState } from 'react'
 import styles from "./component.module.css"
-import MediaListCoverInfo from '../MediaItemCoverInfo2'
-import CardMediaCoverAndDescription from '../CardMediaCoverAndDescription'
-import NavButtons from '../NavButtons'
+import MediaListCoverInfo from '../../MediaItemCoverInfo2'
+import CardMediaCoverAndDescription from '../../CardMediaCoverAndDescription'
+import NavButtons from '../../NavButtons'
 import { ApiAiringMidiaResults, ApiDefaultResult } from '@/app/ts/interfaces/apiAnilistDataInterface'
 import API from "@/api/anilist"
