+++ app/media/[id]/components/PaginateItems/components.module.css
@@ -48,7 +48,7 @@
 
     border-color: var(--brand-color);
     color: var(--brand-color);
-    
+
 }
 
 .container .previous_btn a,
